feat: create dedicated package for shared types
