+++ packages/react/package.json
@@ -14,7 +14,6 @@
   ],
   "main": "./dist/index.js",
   "types": "./dist/index.d.ts",
-  "dependencies": {},
   "peerDependencies": {
     "@fedimint/core-web": "workspace:*",
     "react": "^18.3.1"
